move tests into cart tests
